Quest functionality and forms separated in different file for better visibility
